bradesco: fix intermittent test date
The `buildPayload` unit test of the `bradesco` provider
had a problem, some times the test failed and others it
didn't. The issue was that the dates used on the comparison
were not using the `moment-timezone` library, so if the test
ran late on a day, it would think it is the next day and
the `buildPayload` function which uses `moment-timezone`
would get the correct date on the correct timezone. This
commit fixes this issue by adding the timezone to the test.